Removing repeated code within fpp-to-json tool
